+++ GhettosFirearmSDKv2/Firearms/BoltAssemblies/OpenBolt.cs
@@ -495,7 +495,7 @@ namespace GhettosFirearmSDKv2
             joint.angularZMotion = ConfigurableJointMotion.Locked;
         }
 
-        private float BoltLerp(float startTime, int rpm)
+        private float BoltLerp(float startTime, float rpm)
         {
             float timeThatPassed = Time.time - startTime;
             float timeForOneRound = 60f / rpm;
